[codegen] more consumer fusion (#21521)
Enable consumer fusion to fuse more trailing operations, in particular
to fuse all operations used in layer normalization. The particular
enablement is to allow fusion of consumers of fused operations that are
not the "main" operation into which fusion happens.

This required three modifications:

1. Generate replacement values for any fused operation whose results are
used by operations dominated by the main operation (that will turn into
a loop). This is needed because the consumer fusion finds candidates
based on `insert_slice` operations in the loop that are not produced
unless the replacement values are generated. In the case of layer
normalizaiton, we also need actual replacement values so we can return
them. Spurious values are being cleaned up by canonicalization patterns
in the same pass.
2. Avoid generating replacement values for values used as DPS inits as
those cannot (currently) be fused.
3. Sort consumers to fuse by dominance and first most-dominating
consumer fuse to avoid triangular-use situation that prevents fusion.

This generally makes fusion more aggressive on the consumer side.

Signed-off-by: Alex Zinenko <git@ozinenko.com>
Signed-off-by: Jorn Tuyls <jorn.tuyls@gmail.com>